requirements.txt
requests
pandas
matplotlib
psycopg2-binary
